Update theme and provider to v5
