+++ vite.config.ts
@@ -11,4 +11,7 @@ export default defineConfig({
       transformMixedEsModules: true,
     },
   },
+  resolve: {
+    preserveSymlinks: true,
+  },
 });
